+++ mlpp_features/__init__.py
@@ -1,4 +1,4 @@
-from mlpp_features import utils  # load the accessor
+from mlpp_features import accessors  # load the accessor
 
 from mlpp_features.nwp import *
 from mlpp_features.obs import *
